pyproject.toml
[tool.poetry]
name = "education-drf"
version = "0.1.0"
description = ""
authors = ["DimaSimakov <dimarodeo@gmail.com>"]
readme = "README.md"
packages = [{include = "education_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
pillow = "^10.0.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.8"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"
stripe = "^6.7.0"
python-dotenv = "^1.0.0"
celery = "^5.3.4"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
